Fix sign-extension on char routines

+++ src/http/authFile.c
@@ -387,7 +387,7 @@ MaAcl maParseAcl(MaAuth *auth, cchar *aclStr)
         if (aclStr[0] == '0' && aclStr[1] == 'x') {
             aclStr += 2;
         }
-        for (; isxdigit((int) *aclStr); aclStr++) {
+        for (; isxdigit((uchar) *aclStr); aclStr++) {
             c = (int) tolower((int) *aclStr);
             if ('0' <= c && c <= '9') {
                 acl = (acl * 16) + c - '0';
